+++ packages/atlassian-jira-software/src/main/scripts/02.preinstall.sh
@@ -4,11 +4,12 @@ getent passwd @{appUserName} > /dev/null || useradd -r -m -g @{appGroupName} @{a
 
 # Uninstall or Update => stop service
 if [ "$1" = "0" -o "$1" = "2" ]; then
-	service @{appServiceName} stop
+  service @{appServiceName} stop
 fi
 
 # When we update, make sure we're clean
 if [ "$1" = "2" ]; then
   rm -rf @{destBase}/temp/*
   rm -rf @{destBase}/work/*
+  rm -rf @{destBase}/webapps/*
 fi