+++ packages/atlassian-jira-software/src/main/scripts/04.postinstall.sh
@@ -16,10 +16,5 @@ chown -R root:root @{destConf}
 chmod -R 755 @{destBase}/bin/*.sh
 
 # Link back from /etc
-ln -sf @{destConf}/catalina.policy @{destBase}/conf/catalina.policy
-ln -sf @{destConf}/catalina.properties @{destBase}/conf/catalina.properties
-ln -sf @{destConf}/context.xml @{destBase}/conf/context.xml
-ln -sf @{destConf}/logging.properties @{destBase}/conf/logging.properties
-ln -sf @{destConf}/server.xml @{destBase}/conf/server.xml
-ln -sf @{destConf}/tomcat-users.xml @{destBase}/conf/tomcat-users.xml
-ln -sf @{destConf}/web.xml @{destBase}/conf/web.xml
+rm -rf @{destBase}/conf
+ln -sf @{destConf} @{destBase}/conf