+++ packages/nexus/src/main/scripts/04.postinstall.sh
@@ -1,22 +1,20 @@
 # Initial installation
 if [ "$1" = "1" ]; then
   mkdir -p -m775 @{destBase}/{conf,logs,temp,work}
-  mkdir -p -m775 @{appWorkFolder}/conf
-  chown -R @{appUserName}:@{appGroupName} @{destBase}
-  chown -R @{appUserName}:@{appGroupName} @{appWorkFolder}
-
-  cd /etc/rc.d/init.d
-  ln -sf tomcat @{appServiceName}
 
   chkconfig --add @{appServiceName}
 fi
 
+# Fix user rights
+chown -R @{appUserName}:@{appGroupName} @{destBase}
+chown -R @{appUserName}:@{appGroupName} @{appWorkFolder}
+
 # Workaround for BUG: http://jira.codehaus.org/browse/MRPM-89
 chown -R root:root @{destConf}
 
+# Make scripts executable
+chmod -R 755 @{destBase}/bin/*.sh
+
 # Link back from /etc
-ln -sf @{destConf}/catalina.policy @{destBase}/conf/catalina.policy
-ln -sf @{destConf}/catalina.properties @{destBase}/conf/catalina.properties
-ln -sf @{destConf}/context.xml @{destBase}/conf/context.xml
-ln -sf @{destConf}/server.xml @{destBase}/conf/server.xml
-ln -sf @{destConf}/web.xml @{destBase}/conf/web.xml
+rm -rf @{destBase}/conf
+ln -sf @{destConf} @{destBase}/conf