+++ CHANGES.md
@@ -7,6 +7,7 @@ Version 0.174 (05 December 2015)
 * apache-activemq-5.13.0-bin.tar.gz
 * atlassian-jira-software-7.0.3-jira-7.0.3.tar.gz
 * cmake-3.4.1.tar.gz
+* nexus-2.11.4-01-bundle.zip
 
 
 Version 0.173 (01 December 2015)